Editorial work on the POD.
Switched to a style of listing just the method name in the header.  This makes it
easier to search for and link to.  It also allows for a more expansive explaination
of the interface illustrating what is required and what is optional.

Document that there is a default test name.